fixes #123 - Added clear and copy buttons
Added a button which can be used to clear result area and added copy
buttons which can be used to copy the generated url and result.
Now user will not have to select the entire result or url to copy.
Instead the buttons can be clicked to copy content to clipboard.